server.tests: Split DataProviderServiceTests into small test classes
The previous DataProviderServiceTests contains many test cases and
it became quite complex. It's was becoming harder to read and to add
additional test cases.

This commit splits the class into smaller test classes, each focusing
on a sub-component, such as XY, TimeGraph, DataTree, Table,
Configuration and common part.
